# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Controle Financeiro",
    "version": "1.0.0",
    "description": "API para gerenciamento de finanças pessoais",
    "contact": {
      "name": "Suporte",
      "email": "support@financeiro.app"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Autenticação via cookie de sessão (para aplicação web)"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Token de acesso à API (para integrações externas)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de acesso JWT (para autenticação programática)"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "apiKeyAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/reports/pdf": {
      "get": {
        "summary": "Gera PDF do relatório financeiro",
        "description": "Cria um PDF com os dados financeiros formatados igual à página de relatórios",
        "tags": [
          "Relatórios"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "URL do PDF gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Status da operação"
                    },
                    "downloadUrl": {
                      "type": "string",
                      "description": "URL para download do PDF"
                    },
                    "filename": {
                      "type": "string",
                      "description": "Nome do arquivo PDF"
                    }
                  },
                  "example": {
                    "success": true,
                    "downloadUrl": "/api/reports/download/relatorio-financeiro-2024-05-27.pdf",
                    "filename": "relatorio-financeiro-2024-05-27.pdf"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "404": {
            "description": "Carteira não encontrada"
          },
          "500": {
            "description": "Erro ao gerar PDF"
          }
        }
      }
    },
    "/api/reports/download/{filename}": {
      "get": {
        "summary": "Download do arquivo PDF do relatório",
        "description": "Faz o download do arquivo PDF gerado",
        "tags": [
          "Relatórios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Nome do arquivo PDF para download"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo PDF",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Arquivo não encontrado"
          }
        }
      }
    },
...
  }
}